Move utility files to utils folder
