+++ react/utils/displayButtonTypes.js
